225 upstream commits, 273 files, +23400/-2458. Only five files intersect what this
fork carries, and the two kernel files merged without conflict:

  - exl3_gemm.cu / exl3_gemm_kernel.cuh: upstream reworked both (sliced MGEMM mode,
    CPU-MoE range filtering) and dc1dc23's sm_90+ barrier gate survived intact --
    verified line by line rather than assumed. All three EXL3_SM90_BARRIER guards are
    in place, both group_barrier sites still fall back to grid.sync() when the macro
    is unset, upstream added no new barrier sites, and both cooperative launch paths
    dispatch kernels from the guarded header. Upstream has still not addressed the
    barrier itself, so the gate is still needed.
  - setup.py: the -DEXL3_SM90_BARRIER plumbing survived upstream's rewrite.

The qbench conflicts needed judgement rather than a side. engines.py hunk 1 is
additive both ways -- our in-process quantize path and upstream's w4a4_sim -- and
keeps both. The remaining three are one disagreement: upstream now excludes
token_embd from the storage walk, this fork counts it in its own bucket. Ours
stands, and the reason is written at the point of divergence: the EXL3 embed+head
tax against GGUF is only visible if the embedding is counted, and every figure in
the consuming project's docs/embeddings.md and docs/qbench.md was taken that way, so
adopting the exclusion would silently break comparability with all of them rather
than produce an obvious error.

Upstream's NextN/MTP exclusion *is* adopted, because it is orthogonal to that -- a
draft head is a model feature rather than a format property -- and because leaving
mtp_prefixes computed and unused is a trap for the next reader. Not adopted: the
per_layer_token_embd exclusion, which is embedding-shaped and follows the embedding
policy, and the hc_*/*_conv1d exclusions, which are untested here.

plot.py keeps our slot-reservation fix (reserving all six fixed colours unconditionally
spent 6 of 10 slots and made a project with five or more groups fail to plot at all)
and adopts upstream's GGUF colour index.

+++ exllamav3/exllamav3_ext/cache/q_cache_kernels.cuh
@@ -321,9 +321,9 @@ void quant_cache_paged_kernel
     int base = token_pos * groups_per_token + g0;
     int in_base = in_pos * groups_per_token + g0;
 
-    quant_block_x4<k_bits>(k_in + in_base * 32, k_out + base * k_bits, k_out_scales + base, sh_pack[warp], active, compand_a);
+    quant_block_x4<k_bits>(k_in + (size_t) in_base * 32, k_out + (size_t) base * k_bits, k_out_scales + base, sh_pack[warp], active, compand_a);
     __syncwarp();
-    quant_block_x4<v_bits>(v_in + in_base * 32, v_out + base * v_bits, v_out_scales + base, sh_pack[warp], active, compand_a);
+    quant_block_x4<v_bits>(v_in + (size_t) in_base * 32, v_out + (size_t) base * v_bits, v_out_scales + base, sh_pack[warp], active, compand_a);
 }
 
 #define __(i, j) quant_cache_paged_kernel<i, j>
@@ -392,8 +392,8 @@ void dequant_cache_paged_kernel
             + (token_idx % CQ_PAGE_SIZE) : token_pos;
         int base_out = out_pos * groups_per_token + g0;
 
-        dequant_block_x4<k_bits>(k_in + base * k_bits, k_in_scales + base, k_out + base_out * 32, active, compand_a);
-        dequant_block_x4<v_bits>(v_in + base * v_bits, v_in_scales + base, v_out + base_out * 32, active, compand_a);
+        dequant_block_x4<k_bits>(k_in + (size_t) base * k_bits, k_in_scales + base, k_out + (size_t) base_out * 32, active, compand_a);
+        dequant_block_x4<v_bits>(v_in + (size_t) base * v_bits, v_in_scales + base, v_out + (size_t) base_out * 32, active, compand_a);
 
         chunk_id += d_chunks;
     }
